requirements.txt
tensorflow==1.5
keras==2.1.2
flask==1.1.2
pillow==7.0.0
